tracks time for the selected task and updates it
added a timer to the form and whenever it ticks, it adds the amount of that
tick to the time of the selected task. then it finds the matching label on
the form and updates it with the task's time string.

next steps would be to make the times directly editable. I'm considering
just asking the user to input minutes only since that will avoid the
formatting errors I would frequently get with the old form.